add integration test for exporting via CLI
