+++ tests/conftest.py
@@ -18,6 +18,7 @@ from typing import Dict, List
 
 import ee
 import pytest
+from click.testing import CliRunner
 from geedim import Initialize, MaskedImage
 from geedim.utils import root_path
 
@@ -304,6 +305,29 @@ def generic_masked_images(
         landsat_ndvi_masked_image
     ]
 
+@pytest.fixture
+def runner():
+    """ click runner for command line execution. """
+    return CliRunner()
+
+
+@pytest.fixture
+def region_25ha_file() -> pathlib.Path:
+    """ Path to region_25ha geojson file. """
+    return root_path.joinpath('tests/data/region_25ha.geojson')
+
+
+@pytest.fixture
+def region_100ha_file() -> pathlib.Path:
+    """ Path to region_100ha geojson file. """
+    return root_path.joinpath('tests/data/region_100ha.geojson')
+
+
+@pytest.fixture
+def region_10000ha_file() -> pathlib.Path:
+    """ Path to region_10000ha geojson file. """
+    return root_path.joinpath('tests/data/region_10000ha.geojson')
+
 
 def get_image_std(ee_image: ee.Image, region: Dict, std_scale: float):
     """
@@ -319,3 +343,4 @@ def get_image_std(ee_image: ee.Image, region: Dict, std_scale: float):
         reducer='mean', geometry=region, crs=proj.crs(), scale=std_scale, bestEffort=True, maxPixels=1e6
     )
     return mean_std_image.get('TEST').getInfo()
+
